Use libvirt error handlers and custom exception handler.
The libvirt error handler will ignore all error messages, as these errors
should be raised as exceptions anyways and dealt with that way. This prevents
a deliberately ignored exception from still dumping a libvirt error
to the console (such as during domain create when we would us lookupByName
to ensure there wasn't a domain name conflict).