Add guest_from_installer helper method to Installer class.
The method returns a valid Guest instance built around the installers values,
since the installer instance has all the info we need to decide what the
user wants for a Guest class.

Also add a 'guest_lookup' method to CapabilitiesParser, which allows a one
stop function for requesting if a virt type is available and returning the
appropriate Capabilities objects, with useful error messages.