Add selinux_label option to VirtualDisk
If not overridden by the user, selinux_label will be filled with the label
of the specified storage, or "" if we are creating storage.

If the user specifies a label, it will be used to label the new storage or
relabel the existing storage when 'setup()' is called.

This is deliberately crippled at the moment: even if the user requests it,
we will only change the label for existing, local storage that is marked
readonly (basically CD-ROM devices), since that is the priority case at the
moment for Fedora. The code attempts to be robust in the case of selinux
not being present, so this shouldn't change behavior for other distros.