Add 'clone_policy' property to CloneDesign.
This allows the user to list what metrics the Design will use to decide
what devices should be queued up for cloning. Options are:

CLONE_POLICY_NO_READONLY   = 1
CLONE_POLICY_NO_SHAREABLE  = 2
CLONE_POLICY_NO_EMPTYMEDIA = 3

This allows us to change the default CloneDesign behavior, which is to skip
all of the above.