Fix CPU Pinning check for pinning string ending with comma
When you provided CPU pinning string that was ending with comma an
uncaught exception occured of ValueError about invalid literal for
int() with base 10 because of empty string passed to int() function.
This patch fixes the lines ending with comma, lines like "1," and
corrects this to correct CPU pinning string, ie. "1".