Add --initrd-inject option
For automating Fedora installations using Kickstart, a neat trick is
to put the kickstart file into the initrd.  Then there's no need
to have a separate webserver or NFS mount just so anaconda can
retrieve the file.

Example usage:

virt-install --initrd-injections=/path/to/my.ks -x "ks=file:/my.ks" ...

This functionality really needs to be in virt-install because it's
the code which is downloading and then consuming the initrd.
An open question is whether it makes sense to have higher level
options like:

--kickstart=/path/to/my.ks

which would imply both of the above options.  This would be
Anaconda-specific however.