Check /etc/default/keyboard when finding keyboard layout
virt-install checks the host system to determine the keyboard layout
for virtual machines it creates.
One of the places it checks is /etc/default/console-setup.  In the
next version of Debian, the file to look for is /etc/default/keyboard
[1].

Currently, the Debian package has a patch to virt-install to check
this different path:
-CONSOLE_SETUP_CONF = "/etc/default/console-setup"
+CONSOLE_SETUP_CONF = "/etc/default/keyboard"

This patch will check for 'keyboard' and then 'console-setup'