Add support for an Ubuntu HTTP source when it is a mounted ISO
When the installation source is a mounted ISO (or even cobbler imported ISO)
on a HTTP source, the UbuntDistro is not correctly detected causing the
installation to fail by not detecting it as an Ubuntu source.

This patch fixes the approach to determine whether the source is an Ubuntu
source. Additionally, removes the lookup of daily trees as Ubuntu do not
provide them as done with Debian.