Test test case with many console devices wrt new libvirt
Current libvirt does not support multiple <console> devices in
the XML. The virtinst test suite creates these configurations,
but the second device has always been ignored by libvirt.

The next libvirt (0.9.7) will now support multiple console
devices, but with the restriction that only the first console
may be a clone of the <serial> device. The current virtinst
test suite XML is creating a config where the second console
is being treated as clone, due to lack of any target type.

So with the new libvirt the virtinst test suite is now broken
because the previously ignored XML is now being validated and
shown to be non-compliant.

The trivial fix is to just reverse the order of the <console>
devices in the virtinst test